pyproject.toml
[tool.poetry]
name = "repl_python3_destroy-alzheimer"
version = "0.1.0"
description = ""
authors = ["nightsailor <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
cockroachdb = "^0.3.5"
Flask-SQLAlchemy = "^2.5.1"
python-decouple = "^3.5"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
